update login and register pages asplite the functionality
